Search for a NISRA dataset

Search the NISRA data portal for a dataset. You can search dataset titles either for a keyword or with a regular expression, using a dataset code, or by variables that appear in dataset. You can also specify how recently the dataset must have been updated.

Usage

nisra_search(
  keyword = NULL,
  regex = NULL,
  dataset_code = NULL,
  variables = NULL,
  datefrom = NULL,
  flush_cache = FALSE
)

Arguments

keyword

Text to search for in dataset titles

regex

Regular expression for searching dataset titles

dataset_code

Dataset to find

variables

Variables to search for in datasets

datefrom

Date to search from. Search is limited to datasets updated in the last three months if not specified.

flush_cache

Ignore cached values

Value

A tibble of dataset information matching the search terms. This will include dataset codes, label, frequency, dimensions, and dimensions.

Examples

population_datasets <- nisra_search(keyword = "population")
age_datasets <- nisra_search(variables = "age")